Adds correct build information to Build.get_artifacts
The artifacts returned in Build.get_artifacts all have their builds set to
self. This causes an error retrieving fingerprint info when the artifact
was actually generated by a different build. By retrieving the job and
build numbers for artifacts associated with the job, we can ensure that
each artifact has accurate build information and can be properly validated
on save.